Moved helpers to damage.go & style fixes
